Add timestamp per row and tooltip
